# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bug Tracker API",
        "description": "Dokumentasi REST API Bug Tracker untuk QA, DEV, dan PM.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/bugs/{bugId}/attachments": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "List file attachment bug",
                "description": "Mengambil semua file yang ter-attach pada suatu bug.",
                "operationId": "2089891a07344d6f53e5f75bcd16a655",
                "parameters": [
                    {
                        "name": "bugId",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar attachment berhasil diambil"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload file untuk bug",
                "description": "Meng-upload file (misal screenshot) dan menghubungkannya dengan bug tertentu.",
                "operationId": "77d1b1a2b89ed2db0d54ebf8c7df9a6a",
                "parameters": [
                    {
                        "name": "bugId",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "File yang akan di-upload (max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File berhasil di-upload"
                    },
                    "422": {
                        "description": "Validasi gagal (ukuran/tipe file salah, dsb)"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/attachments/{id}": {
            "delete": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Hapus attachment",
                "description": "Menghapus file attachment dari bug.",
                "operationId": "66e2f863588e3bac0aca3972e89726b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID attachment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attachment berhasil dihapus"
                    },
                    "404": {
                        "description": "Attachment tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari /api/login atau /api/register.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}